refactor telegram update classification modules
