+++ packages/agent-telegram/agent-telegram.cabal
@@ -38,6 +38,8 @@ library
     import:           common-options
     hs-source-dirs:   src
     other-modules:    Agent.Telegram.Classify
+                    , Agent.Telegram.Classify.Media
+                    , Agent.Telegram.Classify.User
                     , Agent.Telegram.Internal.Allowlist
                     , Agent.Telegram.Internal.App
                     , Agent.Telegram.Internal.Launch
