+++ packages/agent-telegram/src/Agent/Telegram/Classify.hs
@@ -25,11 +25,12 @@ module Agent.Telegram.Classify
     ) where
 
 import Agent.Telegram.Types
+import Agent.Telegram.Classify.Media
+import Agent.Telegram.Classify.User
 import Control.Applicative ((<|>))
 import Data.Char (isDigit)
-import Data.List (nubBy)
 import qualified Data.Map.Strict as Map
-import Data.Maybe (fromMaybe, isJust, mapMaybe, maybeToList)
+import Data.Maybe (fromMaybe, isJust, mapMaybe)
 import Data.Set (Set)
 import qualified Data.Set as Set
 import Data.Text (Text)
@@ -666,187 +667,6 @@ classifyMessageLike edited bot sender respondToAllGroupMessages message =
             | otherwise =
                 Text.strip (messageContextPrefix bot message <> rewritten)
 
-hasTelegramMedia :: TelegramMessage -> Bool
-hasTelegramMedia = not . null . messageMediaAttachments
-
-messageMediaAttachments :: TelegramMessage -> [TelegramMedia]
-messageMediaAttachments message =
-    concat
-        [ maybeToList photoAttachment
-        , maybeToList documentAttachment
-        , maybeToList audioAttachment
-        , maybeToList videoAttachment
-        , maybeToList videoNoteAttachment
-        , maybeToList animationAttachment
-        , maybeToList stickerAttachment
-        ]
-  where
-    photoAttachment =
-        case reverse message.messagePhoto of
-            photo : _ ->
-                Just TelegramMedia
-                    { telegramMediaKind = TelegramMediaPhoto
-                    , telegramMediaFile =
-                        Just TelegramFileMedia
-                            { fileMediaFileId = photo.photoFileId
-                            , fileMediaName = Nothing
-                            , fileMediaMimeType = Just "image/jpeg"
-                            , fileMediaFileSize = photo.photoFileSize
-                            , fileMediaDuration = Nothing
-                            }
-                    , telegramMediaDescription = "[Photo]"
-                    }
-            [] -> Nothing
-
-    documentAttachment =
-        fmap
-            (\document -> TelegramMedia
-                { telegramMediaKind = TelegramMediaDocument
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = document.documentFileId
-                        , fileMediaName = document.documentFileName
-                        , fileMediaMimeType = document.documentMimeType
-                        , fileMediaFileSize = document.documentFileSize
-                        , fileMediaDuration = Nothing
-                        }
-                , telegramMediaDescription =
-                    "[Document: "
-                        <> fromMaybe document.documentFileId document.documentFileName
-                        <> "]"
-                })
-            message.messageDocument
-
-    audioAttachment =
-        fmap
-            (\audio -> TelegramMedia
-                { telegramMediaKind = TelegramMediaAudio
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = audio.audioFileId
-                        , fileMediaName = audio.audioFileName
-                        , fileMediaMimeType = audio.audioMimeType
-                        , fileMediaFileSize = audio.audioFileSize
-                        , fileMediaDuration = Just audio.audioDuration
-                        }
-                , telegramMediaDescription =
-                    "[Audio file: "
-                        <> fromMaybe "audio" audio.audioFileName
-                        <> "]"
-                })
-            message.messageAudio
-
-    videoAttachment =
-        fmap
-            (\video -> TelegramMedia
-                { telegramMediaKind = TelegramMediaVideo
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = video.videoFileId
-                        , fileMediaName = video.videoFileName
-                        , fileMediaMimeType = video.videoMimeType
-                        , fileMediaFileSize = video.videoFileSize
-                        , fileMediaDuration = Just video.videoDuration
-                        }
-                , telegramMediaDescription =
-                    "[Video file: "
-                        <> fromMaybe "video" video.videoFileName
-                        <> "]"
-                })
-            message.messageVideo
-
-    videoNoteAttachment =
-        fmap
-            (\note -> TelegramMedia
-                { telegramMediaKind = TelegramMediaVideoNote
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = note.videoNoteFileId
-                        , fileMediaName = Nothing
-                        , fileMediaMimeType = Just "video/mp4"
-                        , fileMediaFileSize = note.videoNoteFileSize
-                        , fileMediaDuration = Just note.videoNoteDuration
-                        }
-                , telegramMediaDescription = "[Video note]"
-                })
-            message.messageVideoNote
-
-    animationAttachment =
-        fmap
-            (\animation -> TelegramMedia
-                { telegramMediaKind = TelegramMediaAnimation
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = animation.animationFileId
-                        , fileMediaName = animation.animationFileName
-                        , fileMediaMimeType = animation.animationMimeType
-                        , fileMediaFileSize = animation.animationFileSize
-                        , fileMediaDuration = Nothing
-                        }
-                , telegramMediaDescription = "[Animation]"
-                })
-            message.messageAnimation
-
-    stickerAttachment =
-        fmap
-            (\sticker -> TelegramMedia
-                { telegramMediaKind = TelegramMediaSticker
-                , telegramMediaFile =
-                    Just TelegramFileMedia
-                        { fileMediaFileId = sticker.stickerFileId
-                        , fileMediaName = Nothing
-                        , fileMediaMimeType = Just "application/octet-stream"
-                        , fileMediaFileSize = sticker.stickerFileSize
-                        , fileMediaDuration = Nothing
-                        }
-                , telegramMediaDescription =
-                    "[Sticker"
-                        <> maybe "" (\emoji -> " " <> emoji) sticker.stickerEmoji
-                        <> "]"
-                })
-            message.messageSticker
-
-messageContentText :: TelegramMessage -> Maybe Text
-messageContentText message =
-    case message.messageText <|> message.messageCaption of
-        Just text -> Just text
-        Nothing
-            | Just _ <- message.messageVoice ->
-                Just "[Voice message]"
-            | Just audio <- message.messageAudio ->
-                Just ("[Audio file: " <> fromMaybe "audio" audio.audioFileName <> "]")
-            | Just document <- message.messageDocument ->
-                Just ("[Document: " <> fromMaybe document.documentFileId document.documentFileName <> "]")
-            | not (null message.messagePhoto) ->
-                Just "[Photo]"
-            | Just video <- message.messageVideo ->
-                Just ("[Video file: " <> fromMaybe "video" video.videoFileName <> "]")
-            | Just _ <- message.messageVideoNote ->
-                Just "[Video note]"
-            | Just _ <- message.messageAnimation ->
-                Just "[Animation]"
-            | Just sticker <- message.messageSticker ->
-                Just
-                    ("[Sticker"
-                        <> maybe "" (\emoji -> " " <> emoji) sticker.stickerEmoji
-                        <> "]")
-            | Just location <- message.messageLocation ->
-                Just
-                    ("[Location: "
-                        <> Text.pack (show location.locationLatitude)
-                        <> ", "
-                        <> Text.pack (show location.locationLongitude)
-                        <> "]")
-            | Just contact <- message.messageContact ->
-                Just ("[Contact: " <> contact.contactFirstName <> "]")
-            | Just venue <- message.messageVenue ->
-                Just ("[Venue: " <> venue.venueTitle <> "]")
-            | Just poll <- message.messagePoll ->
-                Just ("[Poll: " <> poll.pollQuestion <> "]")
-            | Just dice <- message.messageDice ->
-                Just ("[Dice: " <> dice.diceEmoji <> " " <> Text.pack (show dice.diceValue) <> "]")
-            | otherwise -> Nothing
-
 messageContextPrefix :: TelegramUser -> TelegramMessage -> Text
 messageContextPrefix bot message =
     forwarded <> replied <> mentioned
@@ -1062,210 +882,6 @@ mentionLabel bot (snippet, mentionedUser)
   where
     mentionName =
         Text.strip (Text.dropWhile (== '@') (Text.strip snippet))
-
-messageMentionEntities :: TelegramMessage -> [(Text, Maybe TelegramUser)]
-messageMentionEntities message =
-    let text = fromMaybe "" (message.messageText <|> message.messageCaption)
-        entities = message.messageEntities <> message.messageCaptionEntities
-    in
-        [ (utf16Slice entity.entityOffset entity.entityLength text, entity.entityUser)
-        | entity <- entities
-        , entity.entityType == "mention" || entity.entityType == "text_mention"
-        ]
-
-mentionedTelegramUsers :: TelegramMessage -> [TelegramUser]
-mentionedTelegramUsers message =
-    nubBy (\left right -> left.userId == right.userId) $
-        mapMaybe snd (messageMentionEntities message)
-
-utf16Slice :: Int -> Int -> Text -> Text
-utf16Slice offset length text =
-    utf16Take length (utf16Drop offset text)
-
-utf16Drop :: Int -> Text -> Text
-utf16Drop n text
-    | n <= 0 = text
-    | otherwise = case Text.uncons text of
-        Nothing -> ""
-        Just (char, rest)
-            | fromEnum char >= 0x10000 -> utf16Drop (n - 2) rest
-            | otherwise -> utf16Drop (n - 1) rest
-
-utf16Take :: Int -> Text -> Text
-utf16Take n text
-    | n <= 0 = ""
-    | otherwise = case Text.uncons text of
-        Nothing -> ""
-        Just (char, rest)
-            | fromEnum char >= 0x10000 ->
-                if n >= 2
-                    then Text.cons char (utf16Take (n - 2) rest)
-                    else ""
-            | otherwise -> Text.cons char (utf16Take (n - 1) rest)
-
-grantableTelegramUser :: TelegramUser -> Bool
-grantableTelegramUser user =
-    not user.userIsBot
-        && not (isAnonymousAdmin user)
-        && user.userId > 0
-
-stubTelegramUser :: Integer -> TelegramUser
-stubTelegramUser userId =
-    TelegramUser
-        { userId
-        , userIsBot = False
-        , userFirstName = Nothing
-        , userLastName = Nothing
-        , userUsername = Nothing
-        }
-
-data TelegramUserResolution
-    = ResolvedTelegramUser TelegramUser
-    | AmbiguousTelegramUsers [TelegramUser]
-    | UnresolvedTelegramUser Text
-    deriving (Eq, Show)
-
-resolveTelegramUser
-    :: TelegramState
-    -> Integer
-    -> Maybe Integer
-    -> Text
-    -> TelegramUserResolution
-resolveTelegramUser state chatId fallbackUserId rawQuery
-    | Text.null query =
-        case fallbackUserId of
-            Just userId -> resolveById userId
-            Nothing ->
-                UnresolvedTelegramUser
-                    "Reply to their message, mention them, or pass a name, \
-                    \@username, or numeric Telegram user id."
-    | Just userId <- parsePositiveUserId query = resolveById userId
-    | otherwise =
-        uniqueMatches
-            (preferChatLocal
-                (filter (userMatchesQuery query) candidates))
-  where
-    query = normalizeUserQuery rawQuery
-    candidates =
-        nubBy (\left right -> left.userId == right.userId)
-            (Map.elems state.seenTelegramUsers)
-
-    resolveById userId =
-        case Map.lookup userId state.seenTelegramUsers of
-            Just user -> grantResolved user
-            Nothing -> grantResolved (stubTelegramUser userId)
-
-    preferChatLocal matches =
-        let local =
-                filter
-                    (\user ->
-                        maybe
-                            False
-                            (Set.member user.userId)
-                            (Map.lookup chatId state.seenUsersByChat))
-                    matches
-        in if null local then matches else local
-
-    uniqueMatches = \case
-        [user] -> grantResolved user
-        [] ->
-            UnresolvedTelegramUser
-                ("I have not seen anyone matching "
-                    <> rawQuery
-                    <> " in this chat yet. Reply to one of their messages \
-                       \with /allow, mention them, or have them send a \
-                       \message here first.")
-        users -> AmbiguousTelegramUsers users
-
-    grantResolved user
-        | grantableTelegramUser user = ResolvedTelegramUser user
-        | otherwise =
-            UnresolvedTelegramUser
-                "That Telegram account cannot be added to the allowlist."
-
-normalizeUserQuery :: Text -> Text
-normalizeUserQuery raw =
-    let stripped = Text.strip raw
-        withoutAt = fromMaybe stripped (Text.stripPrefix "@" stripped)
-        folded = Text.toCaseFold withoutAt
-    in case Text.stripPrefix "user " folded of
-        Just rest | not (Text.null rest) && Text.all isDigit rest -> rest
-        _ -> folded
-
-parsePositiveUserId :: Text -> Maybe Integer
-parsePositiveUserId text = do
-    userId <- readMaybe (Text.unpack text)
-    if userId > 0 then Just userId else Nothing
-
-userMatchesQuery :: Text -> TelegramUser -> Bool
-userMatchesQuery query user =
-    query == Text.pack (show user.userId)
-        || maybe False ((== query) . Text.toCaseFold) user.userUsername
-        || maybe False ((== query) . Text.toCaseFold . Text.strip) user.userFirstName
-        || maybe False ((== query) . Text.toCaseFold . Text.strip) user.userLastName
-        || (let name = Text.toCaseFold (telegramUserDisplayName user)
-            in not (Text.null name) && name == query)
-
-recordSeenTelegramUsers :: TelegramUpdate -> TelegramState -> TelegramState
-recordSeenTelegramUsers update state =
-    foldl' recordOne state (telegramObservedUsers update)
-  where
-    recordOne current (chatId, user) =
-        current
-            { seenTelegramUsers =
-                Map.insert user.userId user current.seenTelegramUsers
-            , seenUsersByChat =
-                Map.insertWith
-                    Set.union
-                    chatId
-                    (Set.singleton user.userId)
-                    current.seenUsersByChat
-            }
-
-telegramObservedUsers :: TelegramUpdate -> [(Integer, TelegramUser)]
-telegramObservedUsers update =
-    nubBy
-        (\(leftChat, leftUser) (rightChat, rightUser) ->
-            leftChat == rightChat && leftUser.userId == rightUser.userId)
-        (maybe [] usersFromMessage update.updateMessage
-            <> maybe [] usersFromMessage update.updateEditedMessage
-            <> maybe [] usersFromReaction update.updateMessageReaction
-            <> maybe [] usersFromCallback update.updateCallbackQuery
-            <> maybe [] usersFromMembership update.updateMyChatMember)
-
-usersFromMessage :: TelegramMessage -> [(Integer, TelegramUser)]
-usersFromMessage message =
-    let chatId = message.messageChat.telegramChatId
-        people =
-            maybeToList message.messageFrom
-                <> maybeToList (message.messageReplyTo >>= (.messageFrom))
-                <> message.messageNewChatMembers
-                <> mentionedTelegramUsers message
-    in map (\user -> (chatId, user)) people
-
-usersFromReaction :: TelegramMessageReaction -> [(Integer, TelegramUser)]
-usersFromReaction reaction =
-    map
-        (\user -> (reaction.messageReactionChat.telegramChatId, user))
-        (maybeToList reaction.messageReactionUser)
-
-usersFromCallback :: TelegramCallbackQuery -> [(Integer, TelegramUser)]
-usersFromCallback callback =
-    (maybe
-        []
-        (\message -> [(message.messageChat.telegramChatId, callback.callbackQueryFrom)])
-        callback.callbackQueryMessage)
-        <> maybe [] usersFromMessage callback.callbackQueryMessage
-
-usersFromMembership :: TelegramChatMemberUpdated -> [(Integer, TelegramUser)]
-usersFromMembership membership =
-    let chatId = membership.chatMemberUpdatedChat.telegramChatId
-    in map
-        (\user -> (chatId, user))
-        [ membership.chatMemberUpdatedFrom
-        , membership.chatMemberUpdatedOld.chatMemberUser
-        , membership.chatMemberUpdatedNew.chatMemberUser
-        ]
 
 reactionMessageText :: TelegramMessageReaction -> Text
 reactionMessageText reaction
